fixup! feat: split childProcess integration into childProcess and workerThreads integrations
